+++ src/main/java/io/supertokens/storage/postgresql/queries/GeneralQueries.java
@@ -98,6 +98,15 @@ public class GeneralQueries {
                     }
                 }
 
+                if (!doesTableExists(start, Config.getConfig(start).getAccessTokenSigningKeysTable())) {
+                    ProcessState.getInstance(start).addState(ProcessState.PROCESS_STATE.CREATING_NEW_TABLE, null);
+                    try (Connection con = ConnectionPool.getConnection(start);
+                         PreparedStatement pst = con
+                                 .prepareStatement(SessionQueries.getQueryToCreateAccessTokenSigningKeysTable(start))) {
+                        pst.executeUpdate();
+                    }
+                }
+
                 if (!doesTableExists(start, Config.getConfig(start).getSessionInfoTable())) {
                     ProcessState.getInstance(start).addState(ProcessState.PROCESS_STATE.CREATING_NEW_TABLE, null);
                     try (Connection con = ConnectionPool.getConnection(start);
@@ -219,6 +228,7 @@ public class GeneralQueries {
         {
             String DROP_QUERY = "DROP TABLE IF EXISTS " + Config.getConfig(start).getKeyValueTable() + "," +
                     Config.getConfig(start).getUsersTable() + "," +
+                    Config.getConfig(start).getAccessTokenSigningKeysTable() + "," +
                     Config.getConfig(start).getSessionInfoTable() + "," +
                     Config.getConfig(start).getEmailPasswordUsersTable() + ","
                     + Config.getConfig(start).getPasswordResetTokensTable() + "," +
@@ -283,6 +293,16 @@ public class GeneralQueries {
             }
         }
         return null;
+    }
+
+    public static void deleteKeyValue_Transaction(Start start, Connection con, String key)
+            throws SQLException, StorageQueryException {
+        String QUERY = "DELETE FROM " + Config.getConfig(start).getKeyValueTable() + " WHERE name = ?";
+
+        try (PreparedStatement pst = con.prepareStatement(QUERY)) {
+            pst.setString(1, key);
+            pst.executeUpdate();
+        }
     }
 
     public static long getUsersCount(Start start, RECIPE_ID[] includeRecipeIds) throws SQLException {